requirements.txt
factory-boy==2.10.0
Django==2.0.1
django-split-settings==0.2.5
django-redis==4.9.0
djangorestframework==3.7.7
mysqlclient==1.3.10
pytz==2017.3
